database error fix

+++ app.js
@@ -4,13 +4,16 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const hbs = require('express-handlebars');
 const csurf = require('csurf');
-const {Handlebars} = require('express-handlebars');
+const dotenv = require('dotenv');
+
+dotenv.config();
+// machine stuff
+const chalk = require('chalk');
 
 const session = require('express-session');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
 // importing database stuff
-const sequelize = require('./utils/database');
 const db = require('./models/index');
 
 // importing routes
@@ -24,21 +27,34 @@ const brandRoutes = require('./routes/brand');
 // importing custom middlewares
 const localesMiddleWare = require('./middlewares/locales');
 const brandMiddleware = require('./middlewares/brand_confirm');
+const loggerMiddleware = require('./middlewares/logger');
+
+
 
 // important variables
-const port = process.env.PORT || 8000;
+const env = process.env.NODE_ENV;
+const port = process.env.PORT;
 const static_dir = path.join(__dirname , 'public');
-
+const db_type = process.env.DATABASE;
+const session_key = process.env.SESSION_KEY;
 
 const app = express();
 const csrfProtection = csurf();
 
 
 
-// setting up express sessions
-const sessionStore = new SequelizeStore({db : sequelize});
+// setting up express session
+let sessionStore = null;
+if(db_type==='mysql'){
+    sessionStore = new SequelizeStore({db : db.sequelize});
+}else if(db_type=='mongodb'){
+
+}else{
+    console.log(chalk.red('No database defined in env'))
+}
+
 app.use(session({
-    secret:"blablahblahbluhbluh",
+    secret:session_key,
     resave : false,
     saveUninitialized : false,
     store : sessionStore,
@@ -66,6 +82,7 @@ app.use('/static' , express.static(static_dir));
 
 app.use(localesMiddleWare);
 app.use(brandMiddleware);
+app.use(loggerMiddleware);
 
 // defing routes
 app.use('/' , websiteRoutes);
@@ -79,13 +96,29 @@ app.use((req,res,next)=>{
 });
 
 // starting the express server
-db.sequelize.sync({force:false})
+
+if(db_type==='mysql'){
+    db.sequelize.sync({force:false})
     .then(()=>{
         app.listen(port , ()=>{
-            console.log(`Three step launchers server started at port ${port}`);
+            console.log(chalk.yellowBright(`[${env} MODE]`));
+            console.log(chalk.blue(`PORT : ${port}`));
+            console.log(chalk.bgYellow(`DATABASE : ${db_type}`));
+            console.log(chalk.green('Three step launchers server started'));
         });
     })
-    .catch((error)=>console.log(error));
+    .catch((error)=>{
+        console.log(chalk.bgRedBright('ERROR OCCOURED !'))
+        console.log(chalk.red(error));
+        
+    });
+}else if(db_type=='mongodb'){
+
+
+}else{
+    console.log('Database defined in environment can\'t be done by programmer !')
+}
+
